rename package entity to dto
