+++ app/src/main/java/com/sanaa/tudee_assistant/data/local/RoomDatabase.kt
@@ -6,8 +6,8 @@ import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import com.sanaa.tudee_assistant.data.local.dao.CategoryDao
 import com.sanaa.tudee_assistant.data.local.dao.TaskDao
-import com.sanaa.tudee_assistant.data.local.entity.CategoryLocalDto
-import com.sanaa.tudee_assistant.data.local.entity.TaskLocalDto
+import com.sanaa.tudee_assistant.data.local.dto.CategoryLocalDto
+import com.sanaa.tudee_assistant.data.local.dto.TaskLocalDto
 
 @Database(
     entities = [TaskLocalDto::class, CategoryLocalDto::class],
