+++ app/src/main/java/com/sanaa/tudee_assistant/data/local/dao/CategoryDao.kt
@@ -6,7 +6,7 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
-import com.sanaa.tudee_assistant.data.local.entity.CategoryLocalDto
+import com.sanaa.tudee_assistant.data.local.dto.CategoryLocalDto
 import kotlinx.coroutines.flow.Flow
 
 @Dao
